QA SP update 3
QA is now requesting the scope of work and a list of known issues before the kickoff meeting. We are also emphasizing the importance of an acceptance test and clarifying that the developers will create the project boards and milestones with the definition of done. QA will assist and guide the developers through those if needed during the kickoff meeting.

closes